echo-core/requirements.txt
Pillow==10.1.0
pywinauto==0.6.8
psutil==5.9.7
pynput==1.7.6
